+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["caci", "attestation", "cbor", "cbor-ffi", "cose", "crypto", "ffi"]
+members = ["caci", "attestation", "cbor", "cose", "crypto", "ffi"]
 resolver = "2"
