--- cbor-ffi/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "tee-attestation-verification-cbor-ffi"
-version = "1.0.8"
-edition = "2021"
-rust-version = "1.77"
-repository = "https://github.com/microsoft/TEE-Attestation-Verification"
-keywords = ["cbor", "ffi", "confidentiality"]
-description = "Handle-based C ABI for building and inspecting CBOR documents"
-authors = ["Max Tropets", "Chris Jensen"]
-license = "MIT"
-
-[lib]
-crate-type = ["staticlib", "rlib"]
-
-[dependencies]
-cbor = { package = "tee-attestation-verification-cbor", version = "1.0.8", path = "../cbor" }
