+++ ffi/Cargo.toml
@@ -35,6 +35,7 @@ crypto_windows = [
 [dependencies]
 attestation = { package = "tee-attestation-verification-lib", version = "1.0.8", path = "../attestation", default-features = false }
 caci = { package = "tee-attestation-verification-caci", version = "1.0.8", path = "../caci", default-features = false }
+cbor = { package = "tee-attestation-verification-cbor", version = "1.0.8", path = "../cbor" }
 cose = { package = "tee-attestation-verification-cose", version = "1.0.8", path = "../cose", default-features = false }
 crypto = { package = "tee-attestation-verification-crypto", version = "1.0.8", path = "../crypto", default-features = false }
 serde_json = ">=1, <2"
