+++ ffi/src/c_ffi/cbor.rs
@@ -1,23 +1,231 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-//! C ABI entry points.
+//! Handle-based C ABI for building, serializing, parsing and inspecting CBOR
+//! documents.
 //!
-//! Every entry point runs its body under [`std::panic::catch_unwind`], so a
-//! panic is reported as a status code or a null handle rather than unwinding
-//! into a C frame, which would abort the host process.
+//! Consumers use the C++ wrapper in `include/tav/cbor.hpp`, which owns
+//! the handles and so upholds the contract below. The ABI itself is declared
+//! in `include/tav/internal/cbor_abi.h`.
+//!
+//! # Handle ownership
+//!
+//! A handle owns one [`CborValue`] and every child below it. Container
+//! constructors consume the handles they are given and null the caller's
+//! variables. Each handle in a batch must be distinct: a repeated handle is
+//! consumed twice and so freed twice. Callers observing this obtain a tree,
+//! which the implementation assumes without validation.
+//!
+//! # Payload ownership
+//!
+//! Scalars are copied. Byte and text payloads are borrowed: a handle stores
+//! the caller's pointer and length. The `'static` in [`TavCborHandle`] is a
+//! claim the caller upholds, since a C handle has no lifetime to name.
+//! Buffers passed to a constructor or a parse call must remain alive and
+//! unmodified while any handle derived from them is in use.
+//!
+//! # Limits
+//!
+//! Parsing and serialization reject nesting deeper than [`MAX_DEPTH_LIMIT`],
+//! whatever depth the caller asks for.
 
+use std::borrow::Cow;
 use std::os::raw::c_char;
 use std::panic::{catch_unwind, AssertUnwindSafe};
 
 use cbor::{CborValue, Det, Mode, Nondet};
 
-use crate::{
-    as_value, borrow, borrowed, bytes_value, capped, into_handle, is_reserved_simple,
-    keys_are_usable, kind_of, string_value, take, take_all, usable_as_key, TavCborHandle,
-    KIND_INVALID, STATUS_DECODE_FAILED, STATUS_ENCODE_FAILED, STATUS_KEY_NOT_FOUND, STATUS_OK,
-    STATUS_OUT_OF_BOUND, STATUS_TYPE_MISMATCH,
-};
+/// Success.
+pub const STATUS_OK: i32 = 0;
+/// Malformed input, or a panic while parsing.
+pub const STATUS_DECODE_FAILED: i32 = 1;
+/// A map key or tag that is not present.
+pub const STATUS_KEY_NOT_FOUND: i32 = 2;
+/// An index past the end of an array or map.
+pub const STATUS_OUT_OF_BOUND: i32 = 3;
+/// An operation applied to the wrong kind of value.
+pub const STATUS_TYPE_MISMATCH: i32 = 4;
+/// An unencodable value, or a panic while serializing.
+pub const STATUS_ENCODE_FAILED: i32 = 5;
+
+/// A null or otherwise unreadable handle.
+pub const KIND_INVALID: i32 = -1;
+pub const KIND_SIGNED: i32 = 0;
+pub const KIND_BYTES: i32 = 1;
+pub const KIND_STRING: i32 = 2;
+pub const KIND_ARRAY: i32 = 3;
+pub const KIND_MAP: i32 = 4;
+pub const KIND_TAGGED: i32 = 5;
+pub const KIND_SIMPLE: i32 = 6;
+
+/// Ceiling on the depth a caller may request, bounding recursion in the
+/// parser and serializer so that deeply nested input cannot overflow the
+/// stack.
+pub const MAX_DEPTH_LIMIT: usize = 256;
+
+/// A CBOR value behind a C handle.
+///
+/// Byte and text payloads may point into caller memory, which the caller
+/// guarantees outlives this value.
+#[repr(transparent)]
+pub struct TavCborHandle(pub(crate) CborValue<'static>);
+
+/// Move `value` onto the heap and hand the caller an owning handle.
+pub(crate) fn into_handle(value: CborValue<'static>) -> *mut TavCborHandle {
+    Box::into_raw(Box::new(TavCborHandle(value)))
+}
+
+/// Read a handle without taking ownership.
+///
+/// # Safety
+/// `handle` must be null or a live handle.
+pub(crate) unsafe fn as_value<'a>(handle: *const TavCborHandle) -> Option<&'a CborValue<'static>> {
+    unsafe { handle.as_ref() }.map(|h| &h.0)
+}
+
+/// Borrow a child as a handle.
+///
+/// Sound because [`TavCborHandle`] is `repr(transparent)` over [`CborValue`],
+/// so a child's own address serves as its handle.
+pub(crate) fn borrow(value: &CborValue<'static>) -> *const TavCborHandle {
+    (value as *const CborValue<'static>).cast()
+}
+
+/// View caller memory as a slice that outlives this call.
+///
+/// # Safety
+/// `data` must be valid for `len` bytes, and that memory must stay alive and
+/// unmodified for as long as any handle built from it is used.
+pub(crate) unsafe fn borrowed(data: *const u8, len: usize) -> Option<&'static [u8]> {
+    if len == 0 {
+        return Some(&[]);
+    }
+    if data.is_null() {
+        return None;
+    }
+    Some(unsafe { std::slice::from_raw_parts(data, len) })
+}
+
+/// Take ownership of the handle in `slot`, leaving null behind.
+///
+/// # Safety
+/// `slot` must be null or point to a writable handle variable.
+pub(crate) unsafe fn take(slot: *mut *mut TavCborHandle) -> Option<CborValue<'static>> {
+    if slot.is_null() {
+        return None;
+    }
+    let handle = unsafe { *slot };
+    if handle.is_null() {
+        return None;
+    }
+    unsafe { *slot = std::ptr::null_mut() };
+    Some(unsafe { Box::from_raw(handle) }.0)
+}
+
+/// Take ownership of `count` handles, returning those already taken if any
+/// slot is null.
+///
+/// Every slot must hold a distinct handle. A handle appearing twice is taken
+/// twice and so freed twice, which this does not detect.
+///
+/// # Safety
+/// `slots` must be valid for `count` handle variables holding distinct
+/// handles.
+pub(crate) unsafe fn take_all(
+    slots: *mut *mut TavCborHandle,
+    count: usize,
+) -> Option<Vec<CborValue<'static>>> {
+    if count == 0 {
+        return Some(Vec::new());
+    }
+    if slots.is_null() {
+        return None;
+    }
+
+    let mut taken = Vec::with_capacity(count);
+    for i in 0..count {
+        match unsafe { take(slots.add(i)) } {
+            Some(value) => taken.push(value),
+            None => {
+                for (j, value) in taken.into_iter().enumerate() {
+                    unsafe { *slots.add(j) = into_handle(value) };
+                }
+                return None;
+            }
+        }
+    }
+    Some(taken)
+}
+
+/// Build a byte string that borrows `payload`.
+pub(crate) fn bytes_value(payload: &'static [u8]) -> CborValue<'static> {
+    CborValue::ByteString(Cow::Borrowed(payload))
+}
+
+/// Build a text string that borrows `payload`, rejecting invalid UTF-8.
+pub(crate) fn string_value(payload: &'static [u8]) -> Option<CborValue<'static>> {
+    std::str::from_utf8(payload)
+        .ok()
+        .map(|text| CborValue::TextString(Cow::Borrowed(text)))
+}
+
+/// Clamp a caller-supplied depth to [`MAX_DEPTH_LIMIT`].
+pub(crate) fn capped(max_depth: usize) -> usize {
+    max_depth.min(MAX_DEPTH_LIMIT)
+}
+
+/// Report the kind of `value` as one of the `KIND_*` constants.
+pub(crate) fn kind_of(value: &CborValue<'static>) -> i32 {
+    match value {
+        CborValue::Int(_) => KIND_SIGNED,
+        CborValue::ByteString(_) => KIND_BYTES,
+        CborValue::TextString(_) => KIND_STRING,
+        CborValue::Array(_) => KIND_ARRAY,
+        CborValue::Map(_) => KIND_MAP,
+        CborValue::Tagged { .. } => KIND_TAGGED,
+        CborValue::Simple(_) => KIND_SIMPLE,
+    }
+}
+
+/// Whether `value` may be used as a map key.
+///
+/// Containers are excluded, so that every key a map can hold is also a key
+/// the C ABI can look up. Parsing enforces the same rule, so a map reached
+/// through this ABI never holds a key that map_at would refuse.
+pub(crate) fn usable_as_key(value: &CborValue<'static>) -> bool {
+    !matches!(
+        value,
+        CborValue::Array(_) | CborValue::Map(_) | CborValue::Tagged { .. }
+    )
+}
+
+/// Whether every map below `value` keys its entries on something usable.
+///
+/// Recursion is bounded by the depth the parse was capped to.
+pub(crate) fn keys_are_usable(value: &CborValue<'static>) -> bool {
+    match value {
+        CborValue::Array(items) => items.iter().all(keys_are_usable),
+        CborValue::Map(entries) => entries
+            .iter()
+            .all(|(key, item)| usable_as_key(key) && keys_are_usable(item)),
+        CborValue::Tagged { payload, .. } => keys_are_usable(payload),
+        _ => true,
+    }
+}
+
+/// Whether `value` is a simple value RFC 8949 reserves.
+///
+/// The reserved range has no encoding, so a handle holding one could be
+/// inspected but never serialized.
+pub(crate) fn is_reserved_simple(value: u8) -> bool {
+    (24..=31).contains(&value)
+}
+
+// --- C ABI entry points ---
+//
+// Every entry point runs its body under `std::panic::catch_unwind`, so a
+// panic is reported as a status code or a null handle rather than unwinding
+// into a C frame, which would abort the host process.
 
 /// Run `body`, returning a null handle if it panics.
 fn guard_handle(body: impl FnOnce() -> *mut TavCborHandle) -> *mut TavCborHandle {
@@ -33,7 +241,7 @@ fn guard_status(on_panic: i32, body: impl FnOnce() -> i32) -> i32 {
 ///
 /// # Safety
 /// `out` must be null or valid for writing.
-unsafe fn scalar_out_ptr<T: Default>(out: *mut T) -> bool {
+unsafe fn reset_scalar_out<T: Default>(out: *mut T) -> bool {
     if out.is_null() {
         return false;
     }
@@ -45,7 +253,7 @@ unsafe fn scalar_out_ptr<T: Default>(out: *mut T) -> bool {
 ///
 /// # Safety
 /// `out` must be null or valid for writing.
-unsafe fn owned_out_ptr<T>(out: *mut *mut T) -> bool {
+unsafe fn reset_owned_out<T>(out: *mut *mut T) -> bool {
     if out.is_null() {
         return false;
     }
@@ -59,8 +267,8 @@ unsafe fn owned_out_ptr<T>(out: *mut *mut T) -> bool {
 /// `err_ptr` and `err_len` must be null or valid for writing.
 unsafe fn reset_error(err_ptr: *mut *mut u8, err_len: *mut usize) {
     unsafe {
-        owned_out_ptr(err_ptr);
-        scalar_out_ptr(err_len);
+        reset_owned_out(err_ptr);
+        reset_scalar_out(err_len);
     }
 }
 
@@ -260,8 +468,8 @@ unsafe fn serialize<M: Mode>(
     err_len: *mut usize,
 ) -> i32 {
     unsafe { reset_error(err_ptr, err_len) };
-    let out_ok = unsafe { owned_out_ptr(out_ptr) };
-    let len_ok = unsafe { scalar_out_ptr(out_len) };
+    let out_ok = unsafe { reset_owned_out(out_ptr) };
+    let len_ok = unsafe { reset_scalar_out(out_len) };
     let Some(value) = (unsafe { as_value(value) }) else {
         unsafe { set_error("Null CBOR handle", err_ptr, err_len) };
         return STATUS_ENCODE_FAILED;
@@ -341,7 +549,7 @@ unsafe fn parse<M: Mode>(
     err_len: *mut usize,
 ) -> i32 {
     unsafe { reset_error(err_ptr, err_len) };
-    if !unsafe { owned_out_ptr(out_value) } {
+    if !unsafe { reset_owned_out(out_value) } {
         unsafe { set_error("Null output pointer", err_ptr, err_len) };
         return STATUS_DECODE_FAILED;
     }
@@ -677,7 +885,7 @@ pub unsafe extern "C" fn tav_cbor_map_key_at(
     out: *mut *const TavCborHandle,
 ) -> i32 {
     guard_status(STATUS_TYPE_MISMATCH, || unsafe {
-        map_entry_at(value, index, out, true)
+        cbor_map_entry_at(value, index, out, true)
     })
 }
 
@@ -692,11 +900,11 @@ pub unsafe extern "C" fn tav_cbor_map_value_at(
     out: *mut *const TavCborHandle,
 ) -> i32 {
     guard_status(STATUS_TYPE_MISMATCH, || unsafe {
-        map_entry_at(value, index, out, false)
+        cbor_map_entry_at(value, index, out, false)
     })
 }
 
-unsafe fn map_entry_at(
+unsafe fn cbor_map_entry_at(
     value: *const TavCborHandle,
     index: usize,
     out: *mut *const TavCborHandle,
