+++ ffi/tests/c-builder/CMakeLists.txt
@@ -6,10 +6,10 @@ set(CMAKE_CXX_STANDARD 20)
 set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
 get_filename_component(TAV_REPO_ROOT "${CMAKE_CURRENT_SOURCE_DIR}/../../.." ABSOLUTE)
-set(TAV_LIB "${TAV_REPO_ROOT}/target/debug/libtee_attestation_verification_cbor_ffi.a")
+set(TAV_LIB "${TAV_REPO_ROOT}/target/debug/libtee_attestation_verification_ffi.a")
 
 add_custom_target(tav_cbor_abi ALL
-    COMMAND cargo build --manifest-path "${TAV_REPO_ROOT}/cbor-ffi/Cargo.toml"
+    COMMAND cargo build --manifest-path "${TAV_REPO_ROOT}/ffi/Cargo.toml" --features crypto_openssl
     WORKING_DIRECTORY "${TAV_REPO_ROOT}"
     BYPRODUCTS "${TAV_LIB}"
     VERBATIM)
@@ -18,8 +18,8 @@ add_executable(tav-cbor-builder-poc main.cpp cbor_builder.cpp)
 add_dependencies(tav-cbor-builder-poc tav_cbor_abi)
 target_include_directories(tav-cbor-builder-poc PRIVATE
     "${TAV_REPO_ROOT}/ffi/tests/c-consumer/vendor"
-    "${TAV_REPO_ROOT}/cbor-ffi/include")
-target_link_libraries(tav-cbor-builder-poc PRIVATE "${TAV_LIB}" pthread dl m)
+    "${TAV_REPO_ROOT}/ffi/include")
+target_link_libraries(tav-cbor-builder-poc PRIVATE "${TAV_LIB}" ssl crypto pthread dl m)
 target_compile_options(tav-cbor-builder-poc PRIVATE -Wall -Wextra)
 
 enable_testing()
